backend/requirements.txt
# Web Framework
flask==2.2.3
flask-restful==0.3.9
flask-cors==3.0.10
gunicorn==20.1.0

# Database
sqlalchemy==1.4.46
psycopg2-binary==2.9.5
influxdb-client==1.36.1
pymongo==4.3.3
redis==4.5.4

# Messaging
confluent-kafka==2.1.1

# Authentication
pyjwt==2.6.0
passlib==1.7.4

# Data Processing
numpy==1.24.2
pandas==1.5.3
scipy==1.10.1
scikit-learn==1.2.2
statsmodels==0.13.5

# Machine Learning
tensorflow==2.12.0
torch==2.0.0
transformers==4.27.4
ray==2.3.1

# API Clients
alpaca-trade-api==3.0.0
polygon-api-client==1.10.1
alpha-vantage==2.3.1
newsapi-python==0.2.7

# Utilities
python-dotenv==1.0.0
pydantic==1.10.7
requests==2.28.2
pyyaml==6.0
python-dateutil==2.8.2
pytz==2023.3
tqdm==4.65.0
joblib==1.2.0

# Testing
pytest==7.3.1
pytest-cov==4.1.0
pytest-mock==3.10.0

